requirements.txt
apache-airflow==2.7.2            
selenium==4.8.0                  
beautifulsoup4==4.12.2         
psycopg2-binary==2.9.3           
pandas==1.5.3                    
requests==2.28.1                 
